removes error.rs
